Create a new UpdateManager implementation for use with WorkManager lib
